Update README.md
Write a C++ program that identifies and counts words and special names in a textual file. There are two kinds of special names considered, called Type 1 and Type 2 names. Counting is applied on all occurrences of words and specialized names.

A word is defined as a sequence of one or more non-whitespace characters separated by whitespace. The special names of Type 1 and Type 2 are identified by a special character at the beginning of the name. Type 1 names must start by a ‘$’ character, while Type 2 names must start by @ character. All Type 1 and Type 2 names should consist of a letter following the special start character, followed by zero or more letters, digits or underscores. For example, $value, @val9, $num_234_ten, and @num_45 are valid Type 1 and Type 2 names, but $9val, and @_num, $num$ are not.